fix(backend): split installation read config binding
