+++ src/backend/tests/community/di/modules/test_installation_read_config.py
@@ -12,16 +12,19 @@ from agentclaw.community.core.skill_center.installation_read_config import Insta
 from agentclaw.community.core.skill_center.services.bot_capability_state_reader import BotCapabilityStateReader
 from agentclaw.community.core.skill_center.version_resolution_contract import SkillVersionResolverProtocol
 from agentclaw.community.di.modules import config_module
+from agentclaw.community.di.modules.installation_read_config_module import (
+    InstallationReadConfigModule,
+)
 
 
 @pytest.mark.parametrize("environment,expected", [("pre", True), ("prepub", True), ("prod", False), ("gray", False), ("dev", False)])
 def test_environment_selection_and_reader_injection(monkeypatch, environment, expected):
     monkeypatch.setenv("SERVER_ENV", environment)
     monkeypatch.setenv("SC_INSTALLATION_DEFAULT_SYNC_ONLY", "true")
-    monkeypatch.setattr(config_module, "_user_config", lambda: {
+    monkeypatch.setattr(config_module, "read_user_config", lambda: {
         "skill_installation": {"default_sync_only": {"pre": True, "prod": False}}
     })
-    injector = Injector([config_module.ConfigModule()])
+    injector = Injector([config_module.ConfigModule(), InstallationReadConfigModule()])
     repository = Mock()
     for protocol, value in [(CapabilityDesiredStateRepositoryProtocol, repository),
                             (BotRepository, Mock()), (SkillsPoolSkillRepositoryProtocol, Mock()),
@@ -40,23 +43,23 @@ def test_environment_selection_and_reader_injection(monkeypatch, environment, ex
 @pytest.mark.parametrize("user_config", [{}, {"skill_installation": {}}, {"skill_installation": {"default_sync_only": {}}}])
 def test_missing_settings_are_disabled(monkeypatch, user_config):
     monkeypatch.setenv("SC_INSTALLATION_DEFAULT_SYNC_ONLY", "true")
-    monkeypatch.setattr(config_module, "_user_config", lambda: user_config)
-    assert config_module.ConfigModule().installation_read() == InstallationReadConfig()
+    monkeypatch.setattr(config_module, "read_user_config", lambda: user_config)
+    assert InstallationReadConfigModule().installation_read() == InstallationReadConfig()
 
 
 @pytest.mark.parametrize("environment,expected", [("pre", False), ("prod", True)])
 def test_production_switch_does_not_enable_pre(monkeypatch, environment, expected):
     monkeypatch.setenv("SERVER_ENV", environment)
-    monkeypatch.setattr(config_module, "_user_config", lambda: {
+    monkeypatch.setattr(config_module, "read_user_config", lambda: {
         "skill_installation": {"default_sync_only": {"prod": True}}
     })
-    assert config_module.ConfigModule().installation_read().default_sync_only is expected
+    assert InstallationReadConfigModule().installation_read().default_sync_only is expected
 
 
 @pytest.mark.parametrize("block", [None, True, {"typo": True}, {"default_sync_only": None},
     {"default_sync_only": True}, {"default_sync_only": {"pre": "false"}},
     {"default_sync_only": {"prod": 1}}, {"default_sync_only": {"prd": True}}])
 def test_invalid_settings_fail_explicitly(monkeypatch, block):
-    monkeypatch.setattr(config_module, "_user_config", lambda: {"skill_installation": block})
+    monkeypatch.setattr(config_module, "read_user_config", lambda: {"skill_installation": block})
     with pytest.raises(ValueError, match="skill_installation"):
-        config_module.ConfigModule().installation_read()
+        InstallationReadConfigModule().installation_read()
